refactor metric and event loading
