+++ cmd/metrics/loader_dynamic.go
@@ -1,0 +1,5 @@
+package metrics
+
+func (l *DynamicLoader) Load(metricDefinitionOverridePath string, eventDefinitionOverridePath string, selectedMetrics []string, metadata Metadata) ([]MetricDefinition, []GroupDefinition, error) {
+	return nil, nil, nil
+}
